--- Assets/DefaultPrefabObjects.asset
@@ -1,18 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3ad70174b079c2f4ebc7931d3dd1af6f, type: 3}
-  m_Name: DefaultPrefabObjects
-  m_EditorClassIdentifier: 
-  <CollectionId>k__BackingField: 0
-  _prefabs:
-  - {fileID: 1670945988362273293, guid: 30d8b9742272cb94189637bf8982c947, type: 3}
-  - {fileID: -1993588417889989405, guid: 057eb476fd753294faae91f83088ae53, type: 3}
