+++ Assets/Scripts/Player.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: fa56b9c7c58fa1346a512ba658a19be9
-NativeFormatImporter:
+guid: 0c9aa9d75bba471469c44af1617b8644
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
