+++ Assets/Scripts/Ship.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 99c9720ab356a0642a771bea13969a05
+guid: f7299300e6854ae498d4cb7f472ec8c3
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
